% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv-pso.R
\name{hbv_pso}
\alias{hbv_pso}
\title{Title HBV Model with Particle Swarm Optimisation (TUWmodel + hydroPSO)}
\usage{
hbv_pso(prec, airt, ep, area = 1, elev_zones = NULL,
  param = ittr::tuwmodel_params_default, obs, from = NULL, to = NULL,
  warmup = 0, pelev = NULL, telev = NULL, incon = NULL,
  outpath = NULL, hydroPSO_args = list(), FUN_gof = hydroGOF::NSE,
  FUN_gof_args = list(), plotting = FALSE)
}
\arguments{
\item{plotting}{}
}
\description{
Function to
}
