% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv-pso-batch.R
\name{hbv_pso_batch}
\alias{hbv_pso_batch}
\title{Batch-run multiple Particle Swarm Optimizations of the HBV Model}
\usage{
hbv_pso_batch(basedir, ...)
}
\arguments{
\item{basedir}{String, file path to the base directory containing glocal_config.R and directories with input data}

\item{...}{Parameters set here will overwrite those set in glocal_config (but overwritten if also set in localconf)}
}
\value{
A list of the following items:
\enumerate{
\item \code{summary} data frame with id, gof, gov_valid, from, to, from_valid and to_valid for each optimization run
\item \code{results} list with hydro_pso output from all optimization runs, named by id
}
In addition, the output files from each optimization run are written inside the respective data input directory, in a sub-directory named by the identifier of the optimization run.
}
\description{
Batch-run multiple Particle Swarm Optimizations of the HBV Model using \link[hydroPSO]{hydroPSO} and \link[TUWmodel]{TUWmodel}.
}
\details{
Runs multiple Particle Swarm Optimizations of HBV for different sets of input data (prec, airt, ep, obs, area, elevation_zones) and/or parameters.

Each optimization run is given an identification string concatenated from the name of the input data folder, the name of the gof variable and an optional suffix. This identifier is used as folder name for the output files, in plot titles and file names, and in the result list returned by \code{hbv_pso_batch}.
}
\section{Directory layout}{

\code{hbv_pso_batch} assumes you have set up your input data in the following way:
\itemize{
\item \code{base directory} A root directory containing the global parameter config and sub-directories for different sets of input data. The path to this directory is the first argument passed to \code{hbv_pso_batch}
\item \code{input data directories} Sub-directories to \code{base directory} containing sets of input data (prec,airt,ep,obs,area and elevation_zones).
\item \code{Data directory} Optional, sub-directory to \code{input data directories} containing a \code{Data} directory taken from HBV-Light (see "Input data definition" below).
}
}

\section{Parameter definition}{

Parameters for the different optimization runs are defined by simple variable assignment, where the variable name corresponds to the parameter as used in \link[ittr]{hbv_pso}.
They can be set in the following locations, in ascending order of precedence (parameters assigned in any location will be overwritten by re-assignment in subsequent locations):
\enumerate{
\item \code{global_config.R} Default parameter definition for all optimization runs. Located at the root of the base directory
\item \code{arguments to hbv_pso_batch} Argument names represent the parameter name to be set (arguments will be unpacked into the environment). Setting individual parameters nested inside lists (e.g. hydroPSOA_args$control$maxit) is currently not possible, meaning any lists passed are not merged but replace the default value.
\item \code{localconf.R} Located inside any input data directory, parameters defined here overwrite all previous definitions and are local to the respective input data set.
}

In addition to parameters described in \link[ittr]{hbv_pso}, the following parameters unique to \code{hbv_pso_batch} can be set:
\enumerate{
\item \code{suffix} String, optional. Will be appended to the optimization run identifier
\item \code{gof.name} String. Used as part of the optimization run identifier
\item \code{from_validation} Date or date string in default format, start of validation run with best parameter set found during optimization (optional)
\item \code{to_validation} Date or date string in default format, end of validation period
}
}

\section{Input data definition}{

For each data input folder, the following variables have to be defined and contain valid values as defined in \link[ittr]{hbv_pso}: prec, airt, ep, obs, area, elev_zones.
The input data can be assinged in the following ways (in descending order of precedence):
\enumerate{
\item assignment in localconf.R. Add code which assigns the input data to the corresponding variable name, e.g. \code{prec <- read.zoo(fpath, custom_options)}
\item text files inside input data directories. The files must be named as the input time series (with either .csv, .txt or no suffix) and produce a valid zoo object with index when read by \link[zoo]{read.zoo} with default arguments
\item import from HBV-light. Dropping a "Data" directory from HBV-light inside the input data directory will allow \code{hbv_pso_batch} to automatically import the data from the format used by HBV-light.
}
If a given input data set is not defined in localconf, the \code{hbv_pso_batch} will try to load it from a csv file, and then (if unsuccesful) from a HBV-light "Data" directory.
}

