% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hbv_single}
\alias{hbv_single}
\title{Title}
\usage{
hbv_single(
  prec,
  airt,
  ep,
  area,
  param,
  obs,
  warmup,
  incon = NULL,
  FUN_gof = hydroGOF::NSE,
  FUN_gof_args = NULL,
  as_hydromod = TRUE,
  par_fixed = NULL,
  disable_tr = FALSE
)
}
\arguments{
\item{FUN_gof}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
