% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly2daily.R
\name{monthly2daily}
\alias{monthly2daily}
\title{Disaggregate Daily Means by Month-of-the-year into Daily Timeseries}
\usage{
monthly2daily(from, to, values, FUN = zoo::na.spline, ...)
}
\arguments{
\item{from}{Beginning of the interpolated time series, as Date object}

\item{to}{End of the interpolated time series, as Date object}

\item{values}{Vector of daily mean values for each month (length=12)}

\item{FUN}{interpolation function to use, must be either na.approx (linear) or na.spline (spline)}

\item{...}{further arguments passed to approx/spline via na.approx/na.spline}
}
\value{
A daily timeseries (zoo) interpolated from daily mean values for each month of the year
}
\description{
Wrapper around na.approx/na.spline which disaggregates monthly to daily values
by linear interpolation or cubic spline interpolation, respectively.
}
\details{
The values for interpolation are mean daily data grouped by month (i.e. 12 values corresponding to
the calendar month, as produced by hydroTSM::monthlyfunction with FUN=mean).

Monthly input data is fixed to the middle of each month for the interpolation, and values at
the left and right margins are interpolated by extending the monthly datapoints accordingly.
}
\author{
Joschka Thurner, \email{joschka.thurner@th-koeln.de}
}
\keyword{internal}
