% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv-light.R
\name{read_clarea}
\alias{read_clarea}
\title{Extract settings from HBV-Light's CLAREA file}
\usage{
read_clarea(clarea_file)
}
\arguments{
\item{clarea_file}{File path to CLAREA.xml}
}
\value{
List containing the following named items:
\item{zones}{Numeric vector of the elevation zones (m.a.s.l)}
\item{area_elev}{Numeric vector of the eleavtion zone's relative}
\item{area_vegetation}{Matrix describing the relative area of the vegetation zones (vegetation zones in columns))}
}
\description{
Reads elevation zones and their relative area into vectors. Caution: Usage of subcatchments is currently not supported. Contatc the author if you require this functionality.
}
\author{
Joschka Thurner  \email{joschka.thurner@th-koeln.de}
}
