% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv-pso.R
\name{apply_lapse_rate}
\alias{apply_lapse_rate}
\title{Title}
\usage{
apply_lapse_rate(series, elev_zones, elev_ref, lapse_rate, type = "rel",
  neg.tozero = TRUE)
}
\arguments{
\item{neg.tozero}{}
}
\description{
Title
}
