% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv-pso.R
\name{hbv_single}
\alias{hbv_single}
\title{Title}
\usage{
hbv_single(prec, airt, ep, area, param, elev_zones, obs, warmup, pelev = NULL,
  telev = NULL, incon = NULL, FUN_gof = hydroGOF::NSE,
  FUN_gof_args = list(), pcalt_applied = FALSE, tcalt_applied = FALSE,
  gof_only = TRUE)
}
\arguments{
\item{FUN_gof}{}
}
\description{
Title
}
\keyword{internal}
