% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly2daily.R
\name{monthly2daily}
\alias{monthly2daily}
\title{Disaggregate Daily Means by Month-of-the-year into Daily Timeseries}
\usage{
monthly2daily(from, to, values, FUN = zoo::na.spline, ...)
}
\arguments{
\item{from}{Beginning of the interpolated time series, as Date object}

\item{to}{End of the interpolated time series, as Date object}

\item{values}{Vector of daily mean values for each month (length=12)}

\item{FUN}{interpolation function to use, must be either na.approx (linear) or na.spline (spline)}

\item{...}{further arguments passed to approx/spline via na.approx/na.spline}
}
\value{
A daily timeseries (zoo) interpolated from daily mean values for each month of the year
}
\description{
Wrapper around na.approx/na.spline which disaggregates monthly to daily values
by linear interpolation or cubic spline interpolation, respectively.
}
\details{
The values for interpolation are mean daily data grouped by month (i.e. 12 values corresponding to
the calendar month, as produced by hydroTSM::monthlyfunction with FUN=mean).

Monthly input data is fixed to the middle of each month for the interpolation, and values at
the left and right margins are interpolated by extending the monthly datapoints accordingly.
}
\examples{
m2d <- monthly2daily(as.Date("2000-01-01"),as.Date("2001-12-31"),values=c(0:6,5:1),FUN=zoo::na.approx)
tail(m2d)
\dontrun{
# using na.approx directly for comparison
m <- zooreg(rep(c(0:6,5:1),2),as.yearmon("2000-01-01"),freq = 12)
sq <- seq(as.Date(start(m)), as.Date(end(m), frac = 1), by = "day")
zd <- na.approx(m, x = as.Date, xout = sq)

# plot both results
# na.approx is missing values at the right margin, and sets the monthly value
# at the first day of the month
plot(m2d,col="green")
lines(zd,col="red")
legend('topright', c("na.approx","monthly2daily"), col = 2:3, lty = 1)
}
}
\author{
Joschka Thurner, \email{joschka.thurner@th-koeln.de}
}
